[RHP-103] refactor: update environment variable handling and add bot routes
refactor:
- app/asgi.py: format environment variable updates for clarity
- app/beat.py: format environment variable updates for clarity
- app/routes/bot.py: add blueprint for managing bot operations, including launch and scheduling
- app/routes/bot/__init__.py: remove obsolete initialization file
- app/routes/logs.py: update import statement for TaskExec
- app/worker.py: format worker initialization for better readability
