+++ app/routes/bot.py
@@ -2,22 +2,24 @@
 
 import json
 import traceback
+from pathlib import Path
 from typing import TYPE_CHECKING
 
-from quart import Response, jsonify, make_response, request
+from quart import Blueprint, Response, jsonify, make_response, request
 from quart import current_app as app
 
 from utils import (  # noqa: F401
     check_latest,
     reload_module,
 )
-
-from . import bot
-from .task_exec import TaskExec
+from utils.status import TaskExec
 
 if TYPE_CHECKING:
     from celery import Celery
     from flask_sqlalchemy import SQLAlchemy
+
+path_template = str(Path(__file__).parent.resolve().joinpath("templates"))
+bot = Blueprint("bot", __name__, template_folder=path_template)
 
 
 @bot.post("/bot/<id_>/<system>/<typebot>")
@@ -116,29 +118,25 @@ async def periodic_bot(id_: int, system: str, typebot: str) -> Response:
             request_data = await request.data
             request_form = await request.form
             request_json = bytes(await request.json, "utf-8").decode("unicode_escape")
-            data_bot2 = (request_json if isinstance(request_data, bytes) else request_data) or request_form
-
-            data_bot = {}
-            data_bot.update({"schedule": "True"})
+            data_bot = (request_json if isinstance(request_data, bytes) else request_data) or request_form
 
             data_bot: dict[str, str, int, list[str]]
 
             # Check if data_bot is enconded
-            if isinstance(data_bot2, bytes):
-                data_bot2 = data_bot.decode("utf-8")
+            if isinstance(data_bot, bytes):
+                data_bot = data_bot.decode("utf-8")
 
-            if isinstance(data_bot2, str):
+            if isinstance(data_bot, str):
                 if "\\" in data_bot:
-                    data_bot2 = data_bot2.replace("\\", "")
+                    data_bot = data_bot.replace("\\", "")
 
-                if "'" in data_bot2:
-                    data_bot2 = data_bot2.replace("'", "")
+                if "'" in data_bot:
+                    data_bot = data_bot.replace("'", "")
 
-                data_bot2 = json.loads(data_bot2)
-                if not isinstance(data_bot2, dict):
+                data_bot = json.loads(data_bot)
+                if not isinstance(data_bot, dict):
                     raise ValueError("Invalid data_bot format")
 
-            data_bot.update(data_bot2)
             files = await request.files
 
             is_started = await TaskExec.task_exec_schedule(
